+++ src/blocks/iframe/view.php
@@ -121,6 +121,31 @@ function newspack_blocks_get_iframe_html( $mode, $src, $height, $width, $is_full
 			clearInterval(timerId);
 		}
 
+		// Add a listener for dynamic resizing if the iframe supports it.
+		window.addEventListener("message", function(event) {
+			// Reject messages from untrusted origins.
+			if (event.origin !== new URL(iframe.src).origin || iframe.contentWindow !== event.source) {
+				return; 
+			}
+
+			var iframeHeight = 0;
+			if (event.data && event.data.height){
+				if (typeof event.data.height === "number") {
+					iframeHeight = event.data.height;
+				} else if (typeof event.data.height === "string") {
+					iframeHeight = Number(event.data.height);
+				}
+			}
+			if (!isNaN(iframeHeight) && iframeHeight > 0) {
+				// Remove height from the iframe's parent element if needed.
+				if (iframe.parentElement && iframe.parentElement.style.height !== "auto") {
+					iframe.parentElement.style.height = "auto";
+				}
+
+				// Set the new height dynamically.
+				iframe.style.height = iframeHeight + "px";
+			}
+		});
 	</script>
 
 	<?php